changing structure temporarily, fixing code form
